Get servo feedback from Arduino
