+++ ros/wall_e_interfaces/msg/battery/BatteryLevel.msg
@@ -1,0 +1,6 @@
+#
+# WALL-E battery level
+#
+
+# Battery level as a percentage
+int32 level
