+++ ros/wall_e_interfaces/msg/servo/ServoPosition.msg
@@ -1,0 +1,9 @@
+#
+# WALL-E servo position as percentage of servo movement range
+#
+
+# Servo name
+string name
+
+# Servo position (0 to 100% of servo movement range)
+float64 position
