+++ ros/wall_e_interfaces/msg/servo/ServoPositions.msg
@@ -1,0 +1,6 @@
+#
+# WALL-E servo positions as percentages of their movement range
+#
+
+# Servo positions
+wall_e_interfaces/ServoPosition[] servos
